agregando metodos para product en algunos atributos como actualizar la imgde un producto y la lista de sizes de un producto

+++ karmelshoes/src/main/java/com/karmelshoes/persistency/entity/ProductEntity.java
@@ -28,7 +28,7 @@ public class ProductEntity {
     @ElementCollection
     @CollectionTable(name = "product_sizes", joinColumns = @JoinColumn(name = "product_id"))
     @Column(name = "size")
-    private List<Long> sizes;
+    private List<Integer> sizes;
     @Column(name = "color")
     private String color;
     @Column(name = "gender", length = 2)
@@ -36,11 +36,11 @@ public class ProductEntity {
     @Column(name = "img")
     private String img;
 
-    public List<Long> getSizes() {
+    public List<Integer> getSizes() {
         return sizes;
     }
 
-    public void setSizes(List<Long> sizes) {
+    public void setSizes(List<Integer> sizes) {
         this.sizes = sizes;
     }
